backend/requirements.txt
google-cloud-vision==3.4.4
google-cloud-documentai==2.20.1
google-cloud-storage==2.13.0
google-cloud-firestore==2.13.1
google-cloud-bigquery==3.13.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
